v0.2.4 - Fix re-eval with the force failing and more
First, big thing, make bot be able to be communicated with if it ends up in a state where the "stopped" event is set because someone told it to stop, but somehow the stop function didn't continue after that point. This has happened before and left the bot stuck in a state where absolutely nothing happens until I was able to get to it's console again.
Second, make it so when anyone causes a `force_reset` guild re-eval, it first off tells you that this might take a while, and secondly actually waits and lets the currently running guild pinger statemachine timer gear realize it's active state is none and shut itself off properly. This is not an idea solution, and really we should make statetimers be able to stop faster, but this will have to do for now. There might be a way to cancel the `asyncio.sleep` function in particular (what's causing the delay in re-evaluation), so I'll look into that.